+++ src/main/java/com/adminViajes/modelo/repositorio/VehiculoRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.adminViajes.modelo.entidad.Vehiculos;
 
-public interface VehiculoRepositorio extends CrudRepository<Vehiculos, Integer>{
+public interface VehiculoRepositorio extends CrudRepository<Vehiculos, Long> {
 
 }
